requirements.txt
flask
sqlalchemy
psycopg2-binary
pytest
python-dotenv